requirements.txt
asgiref
attrs
Automat
certifi
cffi
charset-normalizer
constantly
cryptography
defusedxml
Django
django-allauth
django-flat-json-widget
django-tinymce
gunicorn
httplib2
hyperlink
idna
incremental
Jinja2
MarkupSafe
oauthlib
Pillow
psycopg2-binary
pycparser
pycrypto
PyJWT
pyparsing
python3-openid
pytz
pyxdg
PyYAML
requests
requests-oauthlib
six
sqlparse
Twisted
typing-extensions
urllib3
zope.interface
attrs==21.2.0
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
constantly==15.1.0
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-flat-json-widget==0.1.3
django-tinymce==3.4.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.4.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
pyxdg==0.27
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
urllib3==1.26.7
wincertstore==0.2
zope.interface==5.4.0
